requirements.txt
paddlepaddle-gpu == 2.2.2
numpy == 1.19.5
pandas == 1.1.5